backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic==0.42.0
firebase-admin==6.6.0
python-dotenv==1.0.1
sse-starlette==2.1.3
pydantic==2.9.2
